SVGEllips: explicitly mark closed path with 'Z'
Circles and ellipses are closed shapes by definition, so we mark them as such explicitly rather than implicitly.
In the FontTools Pen API the endPath marks an open contour, closePath a closed one. There is no automatic closing even if last and move points have same coordinates as in this case