svg_pathops: convert arcs to cubics before converting svg to SkPath
This way we don't pass any arcs to pathops.Path, only bezier curves.
We should thus avoid skia turning arcs to conics and then the need to convert conics into qudratic beziers.

We can then also get rid of the call to convertConicsToQuads at the end of skia_path, since we are certain that we don't pass any arcs or conics as input to Skia.
I will do that in a follow up commit