regenerate requirements.txt and dev-requirements.txt using pip-compile tool
It's clearer to keep separate the library's own requirements from setup.py from the ones needed to develop and run tests on it.
Using pip-compile (from pip-tools package) we can generate a requirements.txt and dev-requirements.txt files by simply declaring the top-level dependencies respectively in setup.py and in dev-requirements.in files.
When the top-level deps change, we can run pip-compile --upgrade setup.py or pip-compile --upgrade dev-requirements.in to update the respective .txt files with the full list of concrete dependencies.