extend SVGPath to support FontTools/UFO pen interface
Currently in nanoemoji when we export an SVGPath to a FontTools OpenType glyph, we first convert it to a pathops.Path just so that we can call its draw() method. But this is not necessary, we can add draw (and drawPoints) method to SVGPath class itself, and call the respective FontTools pen commands.
This also improve interoperability of SVGPath with other libraries that understand the FontTools pen protocol.