tox.ini: usedevelop=true if tox env name contains '-dev' particle
tox -e py39-dev will install in editable mode (for when one is in a rush).
One can also use this as a way to quickly bootstrap a development virtual environment.
E.g. after running the above command, a new venv is created in .tox/py39-dev with all
the pinned requirements installed and the current module installed in editable mode,
ready for editing source files and running directly the pytest command therein