handle tricky paths that throw PathOpsError in SVGPath.might_paint()
Fixes https://github.com/googlefonts/picosvg/issues/192

The test svg is from the NotoColorEmoji region-flags VA.svg (Vatican).

Instead of raising an error in SVGPath.might_paint() when skia-pathops throws an error (for whatever reason, it's a bit of a black box for us), it is better to return True, thus presuming that the path does paint something.

This unblocks us from being able to compile the full set of NotoColorEmoji flags