normalize: check both first significant vector's components, not only x
For normalize() to work as intended, the first vector (with a significant magnitude) should be aligned with Vector(1, 0) -- to eliminate rotation and uniform scaling.
However, the code currently was skipping this transformation when the
first vector's .x component was already == 1.0, thus ignoring the .y
component. But there are infinite vectors with .x == 1.0 that are not
aligned with (1, 0).

Fixes https://github.com/googlefonts/picosvg/issues/249