when applying clip path, only use clip-rule for clips and fill-rule for target shape
clip-rule is only valid within clipPath elements; when we intersect clip paths with the referencing shape, we need to use the shape's fill-rule attribute instead of clip-rule to initialize the pathops.Path on which to perform the intersection.
This fixes an issue with the 'G' (from the motto 'ORDEM E PROGRESSO') of the Brazilian flag in noto-emoji (which is included as a test): here the shape has fill-rule='evenodd' whereas the clipping path has clip-rule='nonzero'; before this fix, we were using the shape's clip-rule (absent, defaulting to 'nonzero').