Fix handling length units for width and height
An error is raised if an SVG has width and height with a length that
includes a unit type e.g. "10px"

Fixed by removing px units since is equal to without units but raise
with message for other unhandled units.

Modified handling edge-case of invalid values to raise instead of
returning None.

Ref: https://www.w3.org/TR/SVG11/types.html#DataTypeLength